[refactor] Admin 패키지 구조 변경(#76)
what : 패지키 구조 변경
why : 패지키 구조 변경
