+++ Backend/src/main/java/com/github/backend/persist/course/repository/querydsl/CourseSearchRepository.java
@@ -1,16 +1,18 @@
 package com.github.backend.persist.course.repository.querydsl;
 
 import com.github.backend.persist.course.Course;
+
+
+import com.github.backend.persist.course.QCourse;
 import com.github.backend.persist.course.repository.querydsl.condition.CourseSearchCondition;
-import com.github.backend.persist.entity.QCourse;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import static com.github.backend.persist.entity.QCourse.course;
 import static org.springframework.util.StringUtils.hasText;
+import static com.github.backend.persist.course.QCourse.course;
 
 @Repository
 @RequiredArgsConstructor
