+++ Backend/src/main/java/com/github/backend/persist/gamer/repository/querydsl/GamerSearchRepository.java
@@ -1,6 +1,7 @@
 package com.github.backend.persist.gamer.repository.querydsl;
 
-import static com.github.backend.persist.entity.QGamer.gamer;
+
+import static com.github.backend.persist.gamer.QGamer.gamer;
 import static org.springframework.util.StringUtils.hasText;
 
 import com.github.backend.persist.gamer.Gamer;
