+++ Backend/src/main/java/com/github/backend/service/admin/impl/AdminService.java
@@ -3,6 +3,7 @@ package com.github.backend.service.admin.impl;
 import com.github.backend.exception.admin.AdminException;
 import com.github.backend.exception.course.CourseException;
 import com.github.backend.exception.gamer.GamerException;
+import com.github.backend.persist.gamer.repository.querydsl.GamerSearchRepository;
 import com.github.backend.persist.member.type.Role;
 import com.github.backend.dto.gamer.SearchCourse;
 import com.github.backend.dto.gamer.SearchGamer;
@@ -14,7 +15,6 @@ import com.github.backend.persist.admin.repository.AdminRepository;
 import com.github.backend.persist.course.repository.CourseRepository;
 import com.github.backend.persist.gamer.repository.GamerRepository;
 import com.github.backend.persist.course.repository.querydsl.CourseSearchRepository;
-import com.github.backend.persist.course.repository.querydsl.GamerSearchRepository;
 import com.github.backend.exception.admin.code.AdminErrorCode;
 import com.github.backend.exception.course.code.CourseErrorCode;
 import com.github.backend.exception.gamer.GamerErrorCode;
@@ -43,7 +43,6 @@ public class AdminService {
      */
     @Transactional
     public Admin registerAdmin(String adminId, String password){
-
 
         validateCreateAdmin(adminId);
