+++ Backend/src/main/java/com/github/backend/service/gamer/impl/GamerService.java
@@ -3,7 +3,7 @@ package com.github.backend.service.gamer.impl;
 import com.github.backend.dto.gamer.SearchGamer;
 import com.github.backend.persist.gamer.Gamer;
 import com.github.backend.persist.gamer.repository.GamerRepository;
-import com.github.backend.persist.course.repository.querydsl.GamerSearchRepository;
+import com.github.backend.persist.gamer.repository.querydsl.GamerSearchRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
