+++ Backend/src/main/java/com/github/backend/web/admin/AdminController.java
@@ -1,4 +1,4 @@
-package com.github.backend.web;
+package com.github.backend.web.admin;
 
 import com.github.backend.dto.course.CourseInfoResponse;
 import com.github.backend.dto.admin.CreateAdmin;
@@ -9,7 +9,6 @@ import com.github.backend.dto.gamer.SearchGamer;
 import com.github.backend.dto.course.UpdateCourse;
 import com.github.backend.dto.gamer.UpdateGamer;
 import com.github.backend.dto.common.Result;
-import com.github.backend.model.dto.*;
 import com.github.backend.persist.admin.Admin;
 import com.github.backend.persist.course.Course;
 import com.github.backend.persist.gamer.Gamer;
